+++ packages/adp-tooling/test/unit/preview/adp-preview.test.ts
@@ -120,21 +120,23 @@ describe('AdaptationProject', () => {
         }
     };
 
-    beforeAll(() => {
-        nock(backend)
-            .get((path) => path.startsWith('/sap/bc/lrep/actions/getcsrftoken/'))
-            .reply(200)
-            .persist(true);
-        nock(backend)
-            .put('/sap/bc/lrep/appdescr_variant_preview/?workspacePath=//')
-            .reply(200, {
-                'my.adaptation': mockMergedDescriptor
-            })
-            .persist(true);
-    });
-
     const logger = new ToolsLogger();
     describe('init', () => {
+        beforeAll(() => {
+            nock(backend)
+                .get((path) => path.startsWith('/sap/bc/lrep/actions/getcsrftoken/'))
+                .reply(200)
+                .persist(true);
+            nock(backend)
+                .put('/sap/bc/lrep/appdescr_variant_preview/?workspacePath=//')
+                .reply(200, {
+                    'my.adaptation': mockMergedDescriptor
+                })
+                .persist(true);
+        });
+        afterAll(() => {
+            nock.cleanAll();
+        });
         test('default (no) config', async () => {
             const adp = new AdpPreview(
                 {
@@ -180,7 +182,37 @@ describe('AdaptationProject', () => {
         });
     });
     describe('sync', () => {
+        let secondCall: boolean = false;
+        beforeAll(() => {
+            nock(backend)
+                .get((path) => path.startsWith('/sap/bc/lrep/actions/getcsrftoken/'))
+                .reply(200)
+                .persist(true);
+            nock(backend)
+                .put('/sap/bc/lrep/appdescr_variant_preview/?workspacePath=//')
+                .reply(200, () => {
+                    if (secondCall) {
+                        return {
+                            'my.adaptation': 'testDescriptor'
+                        };
+                    }
+                    return {
+                        'my.adaptation': mockMergedDescriptor
+                    };
+                })
+                .persist(true);
+        });
+
+        afterAll(() => {
+            nock.cleanAll();
+        });
+
+        afterEach(() => {
+            global.__SAP_UX_MANIFEST_SYNC_REQUIRED__ = false;
+        });
+
         test('updates merged descriptor', async () => {
+            global.__SAP_UX_MANIFEST_SYNC_REQUIRED__ = true;
             const adp = new AdpPreview(
                 {
                     target: {
@@ -203,11 +235,102 @@ describe('AdaptationProject', () => {
             await adp.sync();
             expect(adp.descriptor).toBeDefined();
         });
+
+        test('skip updating the merge descriptor if no manifest changes and descriptor was already fetched', async () => {
+            const adp = new AdpPreview(
+                {
+                    target: {
+                        url: backend
+                    }
+                },
+                mockProject as unknown as ReaderCollection,
+                middlewareUtil,
+                logger
+            );
+
+            mockProject.byGlob.mockResolvedValueOnce([
+                {
+                    getPath: () => '/manifest.appdescr_variant',
+                    getBuffer: () => Buffer.from(descriptorVariant)
+                }
+            ]);
+            await adp.init(JSON.parse(descriptorVariant));
+            (adp as any).mergedDescriptor = undefined;
+            await adp.sync();
+            expect(adp.descriptor).toEqual(mockMergedDescriptor);
+            secondCall = true;
+            await adp.sync();
+            secondCall = false;
+            expect(adp.descriptor).not.toEqual('testDescriptor');
+        });
+
+        test('update descriptor if no manifest changes, but this is first descriptor fetch', async () => {
+            const adp = new AdpPreview(
+                {
+                    target: {
+                        url: backend
+                    }
+                },
+                mockProject as unknown as ReaderCollection,
+                middlewareUtil,
+                logger
+            );
+
+            mockProject.byGlob.mockResolvedValueOnce([
+                {
+                    getPath: () => '/manifest.appdescr_variant',
+                    getBuffer: () => Buffer.from(descriptorVariant)
+                }
+            ]);
+            await adp.init(JSON.parse(descriptorVariant));
+            (adp as any).mergedDescriptor = undefined;
+            await adp.sync();
+            expect(adp.descriptor).toEqual(mockMergedDescriptor);
+        });
+
+        test('update descriptor if descriptor was already fetched, but there are manifest changes', async () => {
+            const adp = new AdpPreview(
+                {
+                    target: {
+                        url: backend
+                    }
+                },
+                mockProject as unknown as ReaderCollection,
+                middlewareUtil,
+                logger
+            );
+
+            mockProject.byGlob.mockResolvedValueOnce([
+                {
+                    getPath: () => '/manifest.appdescr_variant',
+                    getBuffer: () => Buffer.from(descriptorVariant)
+                }
+            ]);
+            await adp.init(JSON.parse(descriptorVariant));
+            (adp as any).mergedDescriptor = undefined;
+            await adp.sync();
+            expect(adp.descriptor).toEqual(mockMergedDescriptor);
+            secondCall = true;
+            global.__SAP_UX_MANIFEST_SYNC_REQUIRED__ = true;
+            await adp.sync();
+            secondCall = false;
+            expect(adp.descriptor).toEqual('testDescriptor');
+        });
     });
     describe('proxy', () => {
         let server: SuperTest<Test>;
         const next = jest.fn().mockImplementation((_req, res) => res.status(200).send());
         beforeAll(async () => {
+            nock(backend)
+                .get((path) => path.startsWith('/sap/bc/lrep/actions/getcsrftoken/'))
+                .reply(200)
+                .persist(true);
+            nock(backend)
+                .put('/sap/bc/lrep/appdescr_variant_preview/?workspacePath=//')
+                .reply(200, {
+                    'my.adaptation': mockMergedDescriptor
+                })
+                .persist(true);
             const adp = new AdpPreview(
                 {
                     target: {
@@ -233,6 +356,10 @@ describe('AdaptationProject', () => {
             app.use((req) => fail(`${req.path} should have been intercepted.`));
 
             server = supertest(app);
+        });
+
+        afterAll(() => {
+            nock.cleanAll();
         });
 
         afterEach(() => {
@@ -321,6 +448,16 @@ describe('AdaptationProject', () => {
     describe('addApis', () => {
         let server: SuperTest<Test>;
         beforeAll(async () => {
+            nock(backend)
+                .get((path) => path.startsWith('/sap/bc/lrep/actions/getcsrftoken/'))
+                .reply(200)
+                .persist(true);
+            nock(backend)
+                .put('/sap/bc/lrep/appdescr_variant_preview/?workspacePath=//')
+                .reply(200, {
+                    'my.adaptation': mockMergedDescriptor
+                })
+                .persist(true);
             const adp = new AdpPreview(
                 {
                     target: {