+++ packages/odata-service-inquirer/src/prompts/prompt-helpers.ts
@@ -1,4 +1,3 @@
-import { isAppStudio } from '@sap-ux/btp-utils';
 import type { ListChoiceOptions } from 'inquirer';
 import { ErrorHandler } from '../error-handler/error-handler';
 import { t } from '../i18n';
@@ -12,12 +11,10 @@ export const errorHandler = new ErrorHandler();
  *
  * @param options - optionally include some of the supported datasource type choices
  * @param options.includeNone - Include the `NONE` option in the datasource type prompt
- * @param options.includeProjectSpecificDest - Include the `projectSpecificDestination` option in the datasource type prompt
  * @returns The datasource type choices
  */
 export function getDatasourceTypeChoices({
-    includeNone = false,
-    includeProjectSpecificDest = false
+    includeNone = false
 }: DatasourceTypePromptOptions = {}): ListChoiceOptions[] {
     const choices: ListChoiceOptions[] = [
         {
@@ -27,16 +24,8 @@ export function getDatasourceTypeChoices({
         {
             name: t('prompts.datasourceType.odataServiceUrlChoiceText'),
             value: DatasourceType.odataServiceUrl
-        },
-        { name: t('prompts.datasourceType.businessHubChoiceText'), value: DatasourceType.businessHub }
+        }
     ];
-
-    if (isAppStudio() && includeProjectSpecificDest) {
-        choices.push({
-            name: t('prompts.datasourceType.projectSpecificDestChoiceText'),
-            value: DatasourceType.projectSpecificDestination
-        });
-    }
 
     choices.push({ name: t('prompts.datasourceType.capProjectChoiceText'), value: DatasourceType.capProject });
     choices.push({ name: t('prompts.datasourceType.metadataFileChoiceText'), value: DatasourceType.metadataFile });