+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.2.16
+
+### Patch Changes
+
+-   Updated dependencies [25b90b5]
+    -   @sap-ux/ui-components@1.21.0
+    -   @sap-ux/ui-prompting@0.3.2
+
 ## 0.2.15
 
 ### Patch Changes