+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/odata-service-inquirer
 
+## 0.6.0
+
+### Minor Changes
+
+-   3734fe8: Adds system prompting to `@sap-ux/odata-service-inquirer`
+
+### Patch Changes
+
+-   Updated dependencies [3734fe8]
+    -   @sap-ux/guided-answers-helper@0.1.0
+    -   @sap-ux/btp-utils@0.16.0
+    -   @sap-ux/axios-extension@1.17.2
+    -   @sap-ux/fiori-generator-shared@0.7.6
+    -   @sap-ux/telemetry@0.5.42
+
 ## 0.5.60
 
 ### Patch Changes