+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui-prompting
 
+## 0.3.2
+
+### Patch Changes
+
+-   Updated dependencies [25b90b5]
+    -   @sap-ux/ui-components@1.21.0
+
 ## 0.3.1
 
 ### Patch Changes